Fix bitfield bugs and return to u8 store backing
return to old

+++ bittorrent/src/peer_comm/peer_connection.rs
@@ -637,7 +637,7 @@ impl<'scope, 'f_store: 'scope> PeerConnection {
                 }
                 let num_pieces = torrent_state.num_pieces();
                 log::info!("[Peer: {}] Have all received", self.peer_id);
-                let bitfield = bitvec::bitvec!(usize, bitvec::order::Msb0; 1; num_pieces);
+                let bitfield = bitvec::bitvec!(u8, bitvec::order::Msb0; 1; num_pieces);
                 torrent_state
                     .piece_selector
                     .peer_bitfield(self.conn_id, bitfield.into_boxed_bitslice());
@@ -661,7 +661,7 @@ impl<'scope, 'f_store: 'scope> PeerConnection {
                 }
                 let num_pieces = torrent_state.num_pieces();
                 log::info!("[Peer: {}] Have None received", self.peer_id);
-                let bitfield = bitvec::bitvec!(usize, bitvec::order::Msb0; 0; num_pieces);
+                let bitfield = bitvec::bitvec!(u8, bitvec::order::Msb0; 0; num_pieces);
                 self.not_interested(false);
                 torrent_state
                     .piece_selector
